+++ md5/md5a.c
@@ -159,7 +159,7 @@ typedef struct Algorithm_t {
 	DIGEST_Destroy *Destroy;
 	DIGEST_Update *Update;
 	DIGEST_End *End;
-	char *(*Data)(const void *, unsigned int, char *);
+	char *(*Data)(const void *, size_t, char *);
 	char *(*Fd)(int, char *);
 } Algorithm_t;
 
@@ -710,7 +710,7 @@ static void
 MDFilter(const Algorithm_t *alg, int tee)
 {
 	DIGEST_CTX *pcontext;
-	unsigned int len;
+	size_t len;
 	unsigned char buffer[BUFSIZ];
 	char buf[HEX_DIGEST_LENGTH];
